MA-7515 CID 29192 (#1 of 1): Resource leak (RESOURCE_LEAK).
leaked_storage: Variable pbuf going out of scope leaks the storage it points to

Coverity 42274 STRING_OVERFLOW fix.
Check interface length before copy it in case it over flow.

CID 42278 String not null terminated.
Append '\0' to the end of the string as read()
won't append it as as string.

This commit merge with below commits:
6b327d1 MA-7986 Coverity 42274 STRING_OVERFLOW fix
1ab7d7c MA-7941 CID 42278 String not null terminated

Change-Id: Id947b0f245b22fe64e04b4ff2e60d696cb8eb942
Signed-off-by: guoyin.chen <guoyin.chen@freescale.com>
Signed-off-by: Wang Haoran <Haoran.Wang@freescale.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/475
Reviewed-by: Wang Haoran <elven.wang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/964